requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pillow==10.0.0
doctr==0.6.2
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.3.1
